// API_Client/client/models/product-variant-item-dto.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Online Shop API
 * REST API for the online shop backend
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



export interface ProductVariantItemDto {
    /**
     * Color-ის ID (წინასწარ /colors-ზე შექმნილი)
     */
    'colorId'?: string;
    /**
     * Size-ის ID (წინასწარ /sizes-ზე შექმნილი)
     */
    'sizeId'?: string;
    /**
     * ამ ვარიანტის მარაგი საწყობში (ცალი)
     */
    'stock': number;
    /**
     * ამ ვარიანტის საკუთარი ფასი — თუ არაა მითითებული, product.price გამოიყენება
     */
    'price'?: string;
}